refactored addon
